notworkingone/ai-flask/requirements.txt
flask
paho-mqtt
numpy==1.26.4
opencv-python-headless==4.9.0.80